extra paranoia: fix 'abs()' warnings by leaving behavior unchanged
- I don't have time for extensive testing of any actual game-behavior
  changes like 'abs()-->fabsf()' would introduce, so we'll leave this
  untouched and fix the warnings by explicit casting instead.
  (... this worked fine so far, so no reason to change it, unless done
  properly.)
- refs #31

+++ src/enemy.c
@@ -312,7 +312,7 @@ CheckIfWayIsFreeOfDroids ( float x1 , float y1 , float x2 , float y2 , int OurLe
   DebugPrintf( 2, "\nint CheckIfWayIsFreeOfDroids (...) : Checking from %d-%d to %d-%d.", (int) x1, (int) y1 , (int) x2, (int) y2 );
   fflush(stdout);
 
-  if ( abs(x1-x2) > abs (y1-y2) ) LargerDistance=fabsf(x1-x2);
+  if ( abs((int)(x1-x2)) > abs ((int)(y1-y2)) ) LargerDistance=fabsf(x1-x2);
   else LargerDistance=fabsf(y1-y2);
 
   Steps=LargerDistance * 4 ;   // We check four times on each map tile...
