+++ src/map.c
@@ -1315,10 +1315,10 @@ MoveLevelDoors (void)
 		  AllEnemys[j].levelnum != CurLevel->levelnum)
 		continue;
 
-	      xdist = abs (AllEnemys[j].pos.x - doorx);
+	      xdist = abs ((int)(AllEnemys[j].pos.x - doorx));
 	      if (xdist < Block_Rect.w)
 		{
-		  ydist = abs (AllEnemys[j].pos.y - doory);
+		  ydist = abs ((int)(AllEnemys[j].pos.y - doory));
 		  if (ydist < Block_Rect.h)
 		    {
 		      dist2 = xdist * xdist + ydist * ydist;
